refactor(runtime): inline remote module loading
