+++ apps/website-new/docs/zh/guide/advanced/runtime-size-optimization.mdx
@@ -143,9 +143,9 @@ optimization: {
 
 | 配置                  | 原始大小 |      减少 |
 | --------------------- | -------: | --------: |
-| 保留远程和共享能力    | 72,974 B |         - |
-| `disableRemote: true` | 58,498 B | 约 19.84% |
-| `disableShared: true` | 64,756 B | 约 11.26% |
+| 保留远程和共享能力    | 72,851 B |         - |
+| `disableRemote: true` | 58,394 B | 约 19.84% |
+| `disableShared: true` | 64,642 B | 约 11.27% |
 
 ### 3. 跑真实业务流程
