--- packages/runtime-core/src/remote-module/disabled.ts
@@ -1,7 +1,0 @@
-export class DisabledRemoteModuleHandler {
-  get(): never {
-    throw new Error(
-      'Remote loading is disabled by experiments.optimization.disableRemote.',
-    );
-  }
-}
